backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
groq>=0.4.0
yt-dlp>=2024.12.0
opencc-python-reimplemented
